src/QTL/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
analytics-python==1.4.0
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autograd==1.3
autoray==0.2.5
backoff==1.10.0
bcrypt==3.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
cycler==0.11.0
dill==0.3.4
fastapi==0.75.0
ffmpy==0.3.0
filelock==3.6.0
fonttools==4.31.2
frozenlist==1.3.0
future==0.18.2
gradio==2.8.14
h11==0.13.0
huggingface-hub==0.4.0
idna==3.3
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.0
linkify-it-py==1.0.3
markdown-it-py==2.0.1
MarkupSafe==2.1.1
matplotlib==3.5.1
mdit-py-plugins==0.3.0
mdurl==0.1.0
monotonic==1.6
mpmath==1.2.1
multidict==6.0.2
networkx==2.7.1
ninja==1.10.2.3
ntlm-auth==1.5.0
numpy==1.22.3
orjson==3.6.7
packaging==21.3
pandas==1.4.1
paramiko==2.10.3
pbr==5.8.1
PennyLane==0.22.1
PennyLane-Lightning==0.22.0
Pillow==9.0.1
ply==3.11
psutil==5.9.0
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pydub==0.25.1
pylatexenc==2.10
PyNaCl==1.5.0
pyparsing==3.0.7
python-constraint==1.4.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
qiskit==0.34.2
qiskit-aer==0.10.3
qiskit-ibmq-provider==0.18.3
qiskit-ignis==0.7.0
qiskit-terra==0.19.2
regex==2022.3.15
requests==2.27.1
requests-ntlm==1.1.0
retworkx==0.11.0
sacremoses==0.0.49
scipy==1.8.0
semantic-version==2.6.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
stevedore==3.5.0
symengine==0.9.2
sympy==1.10.1
tokenizers==0.11.6
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.1
transformers==4.17.0
tweedledum==1.1.1
typing_extensions==4.1.1
uc-micro-py==1.0.1
urllib3==1.26.9
uvicorn==0.17.6
websocket-client==1.3.1
yarl==1.7.2
